backend/requirements.txt
Django==3.2.3
setuptools==70.0.0
djangorestframework==3.12.4
djoser==2.1.0
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
#django-debug-toolbar==4.2.0
django-filter==23.5
django-templated-mail==1.1.1
#djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
#social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
python-decouple==3.8
drf-extra-fields==3.7.0
flake8==7.0.0